Begin BSD (kqueue) implementation
This is a simple proof-of-concept program (not tested) that takes a pid as argv[1] and keeps running until the process with that pid has exited. The program returns 0 when watched process exits, or 1 on event error.

The rest is just a matter of using getppid(2) instead of having the user specify the pid and forking the process specified by the user, killing it when a NOTE_EXIT is encountered.